+++ Content/ContentManager.cs
@@ -1,26 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using Microsoft.Xna.Framework;
+﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using ranch_mayhem_engine.UI;
 using JsonSerializer = Newtonsoft.Json.JsonSerializer;
 
 namespace ranch_mayhem_engine.Content;
 
-public class ContentManager
+public static class ContentManager
 {
-    private readonly Dictionary<string, Texture2D> _contents = [];
-    private readonly Dictionary<string, Dictionary<int, SpriteFont>> _fonts = [];
+    private static readonly Dictionary<string, Texture2D> Contents = [];
+    private static readonly Dictionary<string, Dictionary<int, SpriteFont>> Fonts = [];
 
-    private readonly Dictionary<string, AtlasItem> _sprites = [];
-    private Texture2D _textureAtlas;
-    private Texture2D _textureAtlasAttack;
+    private static readonly Dictionary<string, AtlasItem> Sprites = [];
+    private static Texture2D _textureAtlas;
 
-    public void LoadContent(Microsoft.Xna.Framework.Content.ContentManager content)
+    public static void LoadContentFile(Microsoft.Xna.Framework.Content.ContentManager content, string path)
     {
-        var itemsToLoad = LoadJson<ContentItem>("./content/content.json");
+        var itemsToLoad = LoadJson<ContentItem>(path);
 
         foreach (var item in itemsToLoad)
         {
@@ -35,45 +30,35 @@ public class ContentManager
                 case "unknown":
                 default:
                     Logger.Log(
-                        $"{GetType().FullName}::LoadContent Trying to load content with unknown type, name={item.Name}",
+                        $"{typeof(ContentManager)}::LoadContent Trying to load content with unknown type, name={item.Name}",
                         LogLevel.Warning);
                     break;
             }
         }
 
-        Logger.Log($"Loaded {_contents.Count} sprites");
-        Logger.Log($"Loaded {_fonts.Count} fonts");
+        Logger.Log($"Loaded {Contents.Count} sprites");
+        Logger.Log($"Loaded {Fonts.Count} fonts");
     }
 
-    public void LoadDefaultTextureAtlas(Microsoft.Xna.Framework.Content.ContentManager content)
+    public static void LoadTextureAtlas(Microsoft.Xna.Framework.Content.ContentManager content, string path,
+        string atlasName)
     {
-        var items = LoadJson<AtlasItem>("./content/texture_atlas_definition.json");
+        var items = LoadJson<AtlasItem>(path);
 
         foreach (var atlasItem in items)
         {
-            _sprites.Add(atlasItem.Name, atlasItem);
+            Sprites.Add(atlasItem.Name, atlasItem);
         }
 
-        _textureAtlas = content.Load<Texture2D>("texture_atlas");
+        _textureAtlas = content.Load<Texture2D>(atlasName);
+        Logger.Log($"Loaded {Sprites.Count} atlas sprites");
     }
 
-    public void LoadAttackAtlas(Microsoft.Xna.Framework.Content.ContentManager content)
+    public static Texture2D? GetAtlasSprite(string name)
     {
-        var items = LoadJson<AtlasItem>("./content/texture_atlas_attack_definition.json");
-
-        foreach (var atlasItem in items)
+        if (Sprites.TryGetValue(name, out var item))
         {
-            _sprites.Add(atlasItem.Name, atlasItem);
-        }
-
-        _textureAtlasAttack = content.Load<Texture2D>("texture_atlas_attack");
-    }
-
-    public Texture2D GetSprite(string name)
-    {
-        if (_sprites.TryGetValue(name, out var item))
-        {
-            var extracted = new Texture2D(RanchMayhemEngine.UIManager.GraphicsDevice, item.Width, item.Height);
+            var extracted = new Texture2D(RanchMayhemEngine.UiManager.GraphicsDevice, item.Width, item.Height);
 
             var data = new Color[item.Width * item.Height];
             _textureAtlas.GetData(0, new Rectangle((int)item.Position.X, (int)item.Position.Y, item.Width, item.Height),
@@ -87,35 +72,16 @@ public class ContentManager
         return null;
     }
 
-    public Texture2D GetAttackSprite(string name)
+    private static void LoadSprite(ContentItem item, Microsoft.Xna.Framework.Content.ContentManager content)
     {
-        if (_sprites.TryGetValue(name, out var item))
-        {
-            var extracted = new Texture2D(RanchMayhemEngine.UIManager.GraphicsDevice, item.Width, item.Height);
-
-            var data = new Color[item.Width * item.Height];
-            _textureAtlasAttack.GetData(0,
-                new Rectangle((int)item.Position.X, (int)item.Position.Y, item.Width, item.Height),
-                data, 0, data.Length);
-
-            extracted.SetData(data);
-
-            return extracted;
-        }
-
-        return null;
+        Contents.Add(item.Name, content.Load<Texture2D>(item.Name));
     }
 
-    private void LoadSprite(ContentItem item, Microsoft.Xna.Framework.Content.ContentManager content)
-    {
-        _contents.Add(item.Name, content.Load<Texture2D>(item.Name));
-    }
-
-    private void LoadFont(ContentItem item, Microsoft.Xna.Framework.Content.ContentManager content)
+    private static void LoadFont(ContentItem item, Microsoft.Xna.Framework.Content.ContentManager content)
     {
         if (item.Sizes == null || item.Sizes.Count == 0)
         {
-            Logger.Log($"{GetType().FullName}::LoadFont No sizes defined for font={item.Name}");
+            Logger.Log($"{typeof(ContentManager)}::LoadFont No sizes defined for font={item.Name}");
             return;
         }
 
@@ -123,28 +89,28 @@ public class ContentManager
         {
             var font = content.Load<SpriteFont>($"{item.Name}{size}");
 
-            if (!_fonts.TryGetValue(item.Name, out var fontSizes))
+            if (!Fonts.TryGetValue(item.Name, out var fontSizes))
             {
                 fontSizes = new Dictionary<int, SpriteFont>();
-                _fonts[item.Name] = fontSizes;
+                Fonts[item.Name] = fontSizes;
             }
 
             fontSizes[size] = font;
         }
     }
 
-    public Texture2D GetTexture(string name) => _contents[name];
+    public static Texture2D GetTexture(string name) => Contents[name];
 
-    public int GetClosestSize(string name, int size)
+    private static int GetClosestSize(string name, int size)
     {
         // available: 12, 16, 32
         // want: 20
         // should: 16
 
         var diff = 0;
-        var closestSize = _fonts[name].First().Key;
+        var closestSize = Fonts[name].First().Key;
 
-        foreach (var (key, value) in _fonts[name])
+        foreach (var (key, value) in Fonts[name])
         {
             if (key == size) return key;
 
@@ -163,37 +129,37 @@ public class ContentManager
         return closestSize;
     }
 
-    private bool HasSize(string name, int size)
+    private static bool HasSize(string name, int size)
     {
-        return _fonts[name].ContainsKey(size);
+        return Fonts[name].ContainsKey(size);
     }
 
-    public SpriteFont GetFont(string name, int size)
+    public static SpriteFont GetFont(string name, int size)
     {
         if (!HasSize(name, size))
         {
-            return _fonts[name][GetClosestSize(name, size)];
+            return Fonts[name][GetClosestSize(name, size)];
         }
 
-        return _fonts[name][size];
+        return Fonts[name][size];
     }
 
-    public (SpriteFont font, int size) GetFontWithSize(string name, int size)
+    public static (SpriteFont font, int size) GetFontWithSize(string name, int size)
     {
         if (!HasSize(name, size))
         {
             var closest = GetClosestSize(name, size);
-            return (_fonts[name][closest], closest);
+            return (Fonts[name][closest], closest);
         }
 
-        return (_fonts[name][size], size);
+        return (Fonts[name][size], size);
     }
 
     private static List<T> LoadJson<T>(string path)
     {
         using var file = File.OpenText(path);
         var serializer = new JsonSerializer();
-        var items = (List<T>)serializer.Deserialize(file, typeof(List<T>));
+        var items = (List<T>)serializer.Deserialize(file, typeof(List<T>))!;
         return items;
     }