+++ KeyboardManager.cs
@@ -4,22 +4,22 @@ using ranch_mayhem_engine.UI;
 
 namespace ranch_mayhem_engine;
 
-public class KeyboardManager
+public static class KeyboardManager
 {
-    public bool IsInTextBox { get; set; }
-    private readonly Dictionary<Keys, Page> _keyBindings = new();
+    public static bool IsInTextBox { get; set; }
+    private static readonly Dictionary<Keys, Page> _keyBindings = new();
 
-    public void RegisterBinding(Keys key, Page page)
+    public static void RegisterBinding(Keys key, Page page)
     {
         if (!_keyBindings.TryAdd(key, page))
         {
             Logger.Log(
-                $"{GetType().FullName}::RegisterBinding key={KeyboardInput.GetCharFromKey(key)} already exists and will be ignored.",
+                $"{typeof(KeyboardManager)}::RegisterBinding key={KeyboardInput.GetCharFromKey(key)} already exists and will be ignored.",
                 LogLevel.Warning);
         }
     }
 
-    public void Update()
+    public static void Update()
     {
         if (IsInTextBox) return;