+++ CHANGELOG.md
@@ -1,6 +1,40 @@
 # Change Log
 
 All notable changes to this project will be documented in this file. See [versionize](https://github.com/versionize/versionize) for commit guidelines.
+<a name="3.0.0"></a>
+## [3.0.0](https://www.github.com/t-graski/ranch-mayhem-engine/releases/tag/v3.0.0) (2025-04-07)
+
+### Features
+
+* add border position option ([ade5af7](https://www.github.com/t-graski/ranch-mayhem-engine/commit/ade5af7fe9194d5002508af508eec51538584a63))
+* add functions to retrieve container children ([9dac5b6](https://www.github.com/t-graski/ranch-mayhem-engine/commit/9dac5b6eba3dd47f8c7584f57be2ce2c5e10bad1))
+* add generic UiComponentBuilder ([96ad892](https://www.github.com/t-graski/ranch-mayhem-engine/commit/96ad892939a5cb2da4b225d6cdf938f175eeec21))
+* add hover transition ([e5bdd4f](https://www.github.com/t-graski/ranch-mayhem-engine/commit/e5bdd4f57652d6a301b0c96d37bcd582c1d25e15))
+* add text to box ([7638602](https://www.github.com/t-graski/ranch-mayhem-engine/commit/76386029c0b2b0933c9e50f1af0750383767bdb1))
+* change content location ([3b5888f](https://www.github.com/t-graski/ranch-mayhem-engine/commit/3b5888f401048726659a096f057f67633c79e435))
+* change rendering responsibility to pages ([c253cf8](https://www.github.com/t-graski/ranch-mayhem-engine/commit/c253cf8420b00e7925c643edc3ccea0441a60227))
+* make keyboard -and contentmanager static ([8a56505](https://www.github.com/t-graski/ranch-mayhem-engine/commit/8a56505cef3b3e0eaf07812a5c88e2e8328fe783))
+
+### Bug Fixes
+
+* add monogame content reference attribute ([f7518cc](https://www.github.com/t-graski/ranch-mayhem-engine/commit/f7518ccc23f1e6a10dece57911080872e40bcf9e))
+* add nullable where needed and remove unused imports ([a1c39ae](https://www.github.com/t-graski/ranch-mayhem-engine/commit/a1c39ae738dc7b263fae41a7bc1636f6e3992228))
+* add setBorderColor() ([7f10113](https://www.github.com/t-graski/ranch-mayhem-engine/commit/7f101133c9ca26679d4281783d8c80b79b7b0e26))
+* change csproj ([1d4dbe6](https://www.github.com/t-graski/ranch-mayhem-engine/commit/1d4dbe6c1c93602547e0c5d002116c3dbac34878))
+* change default border color to transparent ([093fc15](https://www.github.com/t-graski/ranch-mayhem-engine/commit/093fc15deeb85c788d2bb0a409bc825da9a89f6a))
+* onclick being invoked if component is not visible ([ced03f6](https://www.github.com/t-graski/ranch-mayhem-engine/commit/ced03f6ac4f0a2047862cd7e8ed85b3ea0b62663))
+* remove draw calls ([942856d](https://www.github.com/t-graski/ranch-mayhem-engine/commit/942856dff0d43ae3a7c299410be322a5a90b14a1))
+* remove off click action ([3e692bf](https://www.github.com/t-graski/ranch-mayhem-engine/commit/3e692bfe89c64cdde2eedbd6034308b1fdb27089))
+* remove unneeded logs ([680795e](https://www.github.com/t-graski/ranch-mayhem-engine/commit/680795e6276adbd1200d85d11a74f4e4ff5668e1))
+* remove unused import ([cc5f0ab](https://www.github.com/t-graski/ranch-mayhem-engine/commit/cc5f0abd9264554c1dfa3b3f16a76cff6bb3f562))
+* scale font to global scale ([534657c](https://www.github.com/t-graski/ranch-mayhem-engine/commit/534657c0fa8f18cbd8b4cf0d88b0215c576e5ab0))
+* version property in csproj ([170832c](https://www.github.com/t-graski/ranch-mayhem-engine/commit/170832c6b39f266e4237b314d91317acb40abfd6))
+
+### Breaking Changes
+
+* change content location ([3b5888f](https://www.github.com/t-graski/ranch-mayhem-engine/commit/3b5888f401048726659a096f057f67633c79e435))
+* change rendering responsibility to pages ([c253cf8](https://www.github.com/t-graski/ranch-mayhem-engine/commit/c253cf8420b00e7925c643edc3ccea0441a60227))
+* make keyboard -and contentmanager static ([8a56505](https://www.github.com/t-graski/ranch-mayhem-engine/commit/8a56505cef3b3e0eaf07812a5c88e2e8328fe783))
 
 <a name="2.0.0"></a>
 ## [2.0.0](https://www.github.com/t-graski/ranch-mayhem-engine/releases/tag/v2.0.0) (2025-03-28)